tools/maps/requirements.txt
Pillow>=9.0.0
requests>=2.28.0
